Fix two more pointer-truncation casts for x64 (x64 runtime fix)
From a C4311/C4312 warning sweep of the x64 build (the class of bug that
crashed au4MakeDitherTable): fix the two dangerous "value -> larger
pointer" reconstructions of genuinely truncated pointers.

- VirtualMem.cpp/.hpp: CVirtualMem::u4AsyncLoadThread took its thread
  parameter as uint32, but CreateThread passes `(void*)this` (64-bit on
  x64), so the class `this` was truncated -> the async page-loader thread
  would deref a garbage pointer. Widen the param to uintptr.
- Water.cpp: the water-effect intensity buffers aligned an _alloca stack
  pointer via `uint32(p + 8) & 0xFFFFFFF8`, truncating the 64-bit stack
  address. Use `(uintptr)(p+8) & ~(uintptr)7` so the high bits survive.

Both compile-clean and warning-free on x64; uintptr==uint32 on Win32 so
behaviour is unchanged there (Win32 build still green).

Remaining C4312 reconstructions (documented for follow-up, not startup-
critical): AudioDaemon/TextOverlay sound&text "handle-as-uint32" APIs,
saveload SetItemData/GetItemData DWORD (needs DWORD_PTR), tpassglobals
loader-notify context. GroffIO 1341/1346 is the intended small-index-in-
pointer-slot (safe). The 573 C4311 are mostly benign ptr->number
truncations (Clut.hpp CLUT indices etc.).

+++ jp2_pc/Source/Lib/Sys/VirtualMem.cpp
@@ -1156,7 +1156,7 @@ bool CVirtualMem::bLoadPageRangeImmediate
 //
 uint32 CVirtualMem::u4AsyncLoadThread
 (
-	uint32	u4_user
+	uintptr	u4_user
 )
 //*************************************
 {
